// WADNR.Web/src/app/shared/generated/model/project-note-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectNoteUpsertRequest { 
    ProjectID?: number;
    Note: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectNoteUpsertRequestForm { 
    ProjectID?: FormControl<number>;
    Note: FormControl<string>;
}

export class ProjectNoteUpsertRequestFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Note = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
                Validators.maxLength(8000),
            ],
        }
    );
}